requirements.txt
kmax==4.9.1rc1
GitPython
click
dotenv
google-genai
langchain
langchain-google-genai
langchain-core
langchain-community
singleton-decorator
pandas
langchain-openai
